fix to typescript
